M123-014 loop entities now have their own Why3 module
Previously, we generated a declaration

  exception L_1

for the exception that is used for early exits of a loop (or just the regular
exit), and this declaration was inserted in the Why3 module for the subprogram
body which contains the loop. in the given TN, there is a loop that is
duplicated on the Why side because of unrolling of part of the loop in the
presence of pragma Loop_Invariant. This made the exception declaration appear
twice, which makes Why3 crash.

Now instead of adding a special mechanism to avoid the duplication of the
exception declaration, we use the generic mechanism of Alfa entities to do
that for us, benefitting from the fact that every loop has an attached Loop
entity in Ada. This loop entity now has its own module to contain the
exception declaration. This changes slightly how the exit statements are
generated in loops, and requires loop entities to be considered for
dependencies between modules.

alfa-definition.adb:
   (Mark): also push/pop the loop entity
   (Mark_Violations_For_All_Scopes): Do nothing for loop entities

gnat2why-decls.adb:
  (Translate_Loop_Entity): Code that generates a module for the loop entity,
  with the exception declaration

gnat2why-driver.adb:
  (Translate_Entity): Dispatch loop entities to (Translate_Loop_Entity)

gnat2why-expr-loops.adb:
  (Transform_Exit_Statement): refer to exception using Module.Exception name
  (Wrap_Loop): refer to exception using Module.Exception name, and do not
  generate exception declaration here

why-atree-sprint.adb:
  (W_Exception_Declaration_Pre_Op): Add a new line, otherwise declarations are
  too close

why-inter.adb:
  (Set_SI_Internal) Depending on a loop exception does not add any other
  dependency
  (Dispatch_Entity) loop exception modules go into "context_in_body"
  (To_Why_Id) exception names are capitalized in Why3