[M123-033] Do not depend on order of definition of expression functions
Currently, the order of appearance of expression function definitions in a unit
make a difference for provability. This patch removes this dependency on the
order of definition. We reuse for that the "completion" mechanism already in
place for expression function theories, and we compute a closure of all the
entities on which an expression function depends to include all axioms for
these entities in a new "closure" theory added as a completion.

* gnat2why-driver.adb
(Complete_Entity_Translation): New procedure to complete the translation of
  an entity, after all entities from the same unit (spec or body) have been
  translated.
(Translate_CUnit): For expression functions in other unit specs, use the most
  precise completion theory (the "closure" one). Call the new procedure to
  complete the definition of expression functions.
* gnat2why-nodes.adb|s
(Add_To_Graph): New procedure to add an edge in a directed graph.
(Get_Graph_Closure): New procedure to get the closure of a node in a directed
  graph.
(Node_Graphs): New type for directed graphs of nodes (a map of nodes to node
  sets).
* gnat2why-subprograms.adb|s
(Complete_Expression_Function_Body_Translation): New procedure to generate
  a "closure" theory that includes all available axioms for a given expression
  function, which is then added as a completion of the theory for this
  expression function.
(Translate_Expression_Function_Body): Factor out the magic suffix. Pass the
  entity to the Close_Theory procedure so that dependencies are computed.
* gnat2why.ads
Document the generation of theories for expression functions.
* why-gen-names.adb|s
(WNE_Expr_Fun_Axiom, WNE_Expr_Fun_Closure): New suffixes for the completion
  theories for expression functions.
* why-inter.adb|s
(Close_Theory): Change procedure signature to pass in an entity defined by
  the current theory, and a boolean that controls whether all the dependencies
  of the given entity should be added for inclusion.
(Entity_Dependencies): New graph of dependencies between entities, both
  filled and used by Close_Theory.