[M123-033] Make contracts of funs called in contracts available for proofs
Currently, the axiom defining the contract of a function F is not available
for the proof of subprogram S if F is not called in S. This leads to unprovable
VCs when S contains calls to subprogram whose contract calls F. This change
ensures that these axioms are available, by creating for every subprogram
a "closure" theory that completes the base theory defining the Why program
function for the subprogram. This "closure" is attached as a completion of
the base theory, so it is included everywhere the base theory is included.
Note that it extends the previous same mechanism for expression functions.

* gnat2why-driver.adb
(Complete_Entity_Translation): Generate completion theory for all subprograms.
(Translate_CUnit): Use "closure" completion for all subprograms in other units.
* gnat2why-subprograms.adb|s
(Complete_Subprogram_Spec_Translation): Replace previous procedure applying
  only to expression functions.
* gnat2why.ads
Document new generation of "closure" theories.