[M121-039] Avoid cloning standard theory that copies axiom Assoc
The Assoc axiom for associativity of an arithmetic operation should not be
passed to alt-ergo, as it leads the prover into a large number of useless
instantiations. Alt-Ergo driver removes such axioms from the standard theories,
but we were cloning int.Exponentiation in _gnatprove_standard.mlw, which was
cloning Assoc as a result. This is corrected by cloning the standard
real.PowerInt instead of introducing our own, and the standard takes care of
not introducing a clone of Assoc (through a substitution of goals when cloning
int.Exponentiation).