[M124-005] Make it possible to define theries for containers in Why type files
The theories for formal container instantiations will be defined in Why type
files. This makes it necessary that they don't include the completions of the
theories they include, because these completions are defined in the context
files, on which type files cannot depend.

* why-inter.adb
(Add_Use_For_Entity, Close_Theory): Add parameter With_Completion for deciding
  when the completion theories should be used.