[M204-015] Do not generate theory for type not in Alfa, and use name __private
New clone was using the actual name of a type not in Alfa, while in this case
the special "__private" type should be used. Remove the theory creates in
some cases, as well as uses of it.

* gnat2why-decls.adb
(Translate_Variable): Only add a use-clause for type in Alfa.
* gnat2why-driver.adb
(Translate_Entity): Do not generate theory for type not in Alfa.
* gnat2why-types.adb
(Ident_Of_Ada_Type): Use name "__private" for type not in Alfa.