[M219-019] flow analysis - subprogram calls
First stab at dealing with globals, but I am not happy with it. We have
an assumption in the graph that each vertex has a unique key. So far I
have used Flow_Id as keys because each vertex in the CFG can be mapped
reasonably neatly to a node in the AST. Except for globals :(