[M219-019] flow analysis - subprogram calls
Fix issue with visible variables. We now look at the individual
N_Object_Declaration instead of going through the Entity list of the
subprogram.

At least now variables and loop parameters are more consistent: We create
the 'initial and 'final vertices when we encounter them in the AST, not
before.