[LC19-003] Make no_split the default mode, except when --limit-line is used
To completely avoid the risk of combinatorial explosion when generating VCs,
we make no_split the default generation, unless --limit-line is used (typically
interactively). In that case, the default proof mode is then_split, so that
a failure to prove a VC gives a program path that can be shown to the user.
"then_split" is the new name for previous option "normal".

* configuration.adb|s
(Read_Command_Line): Use new default.
* gnatprove.adb
Rename Normal to Then_Split, and do not provide unique default value.