[M227-029] flow analysis - records and arrays
Change Flow_Id structure to support arbitrarily nested record
fields. The idea is that we have Node point to the entity and then
the Component list will contain each component we have selected.
This is possibly a bit heavy-handed, but at least its simple and
checking for equality, etc. will be much easier.