[M402-006] Correct transformation of complex membership tests
Membership tests in postconditions were not expanded in "or else"
combination of individual membership tests. We use this opportunity to
remove the expansion of complex membership tests in SPARK expansion
(which goes in the direction of removing this expansion phase) and deal
with it directly in gnat2why.

* spark/spark_definition.adb
(Mark): Accept complex membership tests.
* why/gnat2why-expr.adb
(Transform_Membership_Expression): Deal with complex membership tests.
* why/why-gen-expr.adb|s
(New_And_Expr): Add correct code for domain EW_Prog and EW_Term.
(New_Or_Expr): Add version that accepts an array instead of 2 elements.