[M410-035] Make value of constants available for proof of clients
Constants defined in a package spec should be available for proof
of clients of the package, whether the constant is fully defined in
the private part (deferred constant) or in the public part. The
previous solution consisting in hiding the definition of deferred
constant from clients has been recognized as not suitable. Instead,
a user should define a parameterless function with a postcondition
to achieve some isolation.

We also use this opportunity to improve the provability of client
code using constants (whether public or deferred) by defining a
closure theory for constants, like it was done for functions under
M123-033.

* why/gnat2why-decls.adb|s
(Complete_Constant_Translation): New procedure to define a closure
completion theory for constants.
(Translate_Constant): Pass parameters to Close_Theory to create a
  graph of dependencies for constants.
* why/gnat2why-driver.adb
(Complete_Entity_Translation): Call suitable completion for constants.
(Translate_CUnit): Create completion for constants from otehr units.
* why/gnat2why.ads
Update documentation of translation phase.
* why/why-gen-names.adb|s
Define suitable named for theories attached to constants.