[M416-016] Do not analyze generic units at all
An error was raised during the (useless) analysis of a generic unit,
because some overloaded entities had the same name, because the SPARK
expansion phase ensuring that this cannot happen was not executed on a
generic unit. Now skip altogether the analysis of generic units.

* gnat2why-driver.adb
(GNAT_To_Why): Skip analysis of generic units.
* why-inter.adb
(Add_Completion): Add assertions to detect potential infinite loop.
  Remove previous patch.