[M307-046] flow analysis - antialiasing
Fix hideous bug in untangling assignment targets. If you
had a specifically short construct like R.A (12) this did not
work. Furthermore if we dealt with an array in a record we did
not correctly include the array as a used variable.

I did this mostly be re-writing and extensively commenting the
existing procedures.