M422-010 do not crash when -u is passed for a generic unit
Because of the correction for M416-016, which disables handling of
generic units completely, no .mlw files were generated at all for
generic units. This is a problem at the gnatprove/gprbuild level,
because one cannot determine whether gnat2why has been run on that unit.
We fix this issue here by creating an empty .mlw file in this case.

gnat2why-driver.adb:
   (Touch_Main_File) new function to create empty main Why file
   (GNAT2Why) call (Touch_Main_File) when stopping because of generic
     units