M423-009 fix problem with naming scheme
Up to now, an assumption of gnat2why was that the body and the spec of
a unit have the same file name. This is not necessarily true and this
patch removes this assumption.

Basically, we now use the "spec" file name everywhere, except for the
"main" Why3 output file. This is needed because the builder usually uses
the body name.

This works because the body is not referenced by other Why3 files.

gnat2why-nodes.adb
  (Spec_File_Name_Without_Suffix)
  (Body_File_Name_Without_Suffix)
    Convenience functions to compute body and spec file names

gnat2why-driver.adb:
  (GNAT_To_Why): Use the body file name for the "main" file and for
  the ".alfa" file

why-inter.adb
  (Init_Why_Files) New variant that takes into account the potentially
    different name for the body
  (Add_Use_For_Entity) call Spec_File_Name_Without_Suffix instead of
    simply File_Name_Without_Suffix