[M411-041, M415-022, M422-026] SPARK_Mode pragma and forward references
This adds the support for SPARK_Mode pragma in gnatprove, and it also
authorizes foward references to suprograms not yet declared (in aspects),
which was previously forbidden. This is done together, as both have an
impact on how scopes are treated in gnat2why, and this was more efficient
to deal with both issues at once.

All entities are now marked recursively, which was done previously only
for types. So a subprogram may refer to other subprograms/types/variables
in its contract, and those will be marked before this one. A stack of
entities is maintained to avoid infinite loops. When a toplevel subprogram
body is analyzed, all violations are reported on this entity, so that
all local entities are translated to Why, or none is. This should match
the future behavior of SPARK_Mode which cannot be applied to local
entities inside a subprogram.

* flow/flow.adb, why/gnat2why-decls.adb, why/gnat2why-driver.adb,
  why/gnat2why-subprograms.adb, why/gnat2why-types.adb, why/why-inter.adb,
  why/gnat2why-nodes.adb
Adapt to renaming of Body_In_SPARK to Subprogram_Body_In_SPARK, and In_SPARK
to Entity_In_SPARK.

* spark/spark_definition.adb|s
Restructure traversal of AST to mark entities recursively, through the new
Mark_Entity subprogram.

* spark/spark_util.adb|s
(Is_Toplevel_Entity): New function returns True for toplevel entities.

* why/gnat2why-expr.adb
(Transform_Statement_Or_Declaration): Ignore pragma SPARK_Mode.