[M323-005] Allow recursive calls in contracts
With the recent change in SPARK_Definition, functions which call themselves
in their contract are not in SPARK. There was however a problem in their
translation, which caused gnatprove to crash. Now fixed.

* why/gnat2why-decls.adb
(Is_Mutable_In_Why): Take into accout subprogram entities.
* why/gnat2why-expr.adb
(Transform_Expr): Call Transform_Identifier to translate callee, instead of
  directly To_Why_Id, so that the mapping in Params is taken into account.
* why/gnat2why-subprograms.adb
(Translate_Subprogram_Spec): Add the function name to the name map, so that
  recursive calls in the contract refer to the local name.