[M411-041] Correct severity of message for SPARK violations
In the 'check' mode of GNATprove, the same errors/warnings should be issued
as in other GNATprove modes. A warning was issued instead of an error, and
an info instead of a warning, due to leftover code from the 'detect' mode
(now removed). Now fixed.

* spark_definition.adb
(Complete_Error_Msg): Issue message with proper severity.