M519-001 fix incorrect function return value for derived types
Previously, we computed the return value of a function from the
identifier. In some cases, the identifier does not have the same return
type as the function entity, e.g. when the call is a primitive operation
of a derived type. We now look at the entity to determine which return
type we can expect from the call.

On the ACATS example which showed this regression, this results in a
conversion that is inserted in addition. I suspect this function was
previously not in SPARK, which explains why we only see the problem now.

gnat2why-expr.adb:
   (Transform_Expr): correct computation of "current type" of the Why
     expression by looking at the entity rather than the identifier of a
     function