[M502-045] Volatile State
Updated rules to allow External, Input_Only, Output_Only aspects
to be specified on type declarations. This allows the
introduction of a rule that an object specified as external
can only be passed as an actual parameter in a subprogram call
whose formal parameter is of a type specified as external.
This ensures that it is known that when analysing a subprogram
that a parameter is external.