[M522-010] Mark violation on current entity as well as toplevel entity
The rewrite of SPARK_Definition lead to violations being only marked on
toplevel entities. This does not work for local entities which do not
have a corresponding toplevel declaration, but may be nonetheless used
in different toplevel entities. For example, class-wide types. The
correct thing to do is mark the violation on both the toplevel entity
and the current one.

* gnat2why/spark/spark_definition.adb
(Current_Toplevel_Entity): Renaming of previous Current_Entity.
(Current_Entity): Return current entity if any.
(Mark_Violation): Mark violation on both toplevel entity and current one.