LB12-015 protect defining axioms for expression functions
For e.g. the identity function of some discrete type, written as
expression function, we would generate the following axiom:

  forall x : int. to_int(f(x)) = x

but this is incorrect. We now generate the following axiom:

  forall x : int. in_range(x) -> to_int(f(x)) = x

that is we add the in_range predicates for all needed parameters as
precondition of the axiom

gnat2why-subprograms.adb
   (Compute_Guard_Formula): compute the needed guards for all parameters
   (Translate_Expression_Function_Body): add the guard formula as
   precondition of the defining axiom