Ticket: [523-041]
Description: Added text to indicate that nonreturning procedures may
be called but there is then a proof obligation that this is calling
code can never be executed. Added text to indicate that none of the
new aspects are allowed on subprogram renaming declarations. Added text
to describe the anti-aliasing checks to be performed in Release 1.