[K222-010] apply GNATprove to tokeneer
Now that we have Refined_State, Refined_Global and
Refined_Depends we no longer need to expose the
consituents of an abstract state in the visible part
of the specs. Started moving things back to the
bodies. Currently blocked by M705-002. Will resume
again as soon as that is resolved.