[M704-35, K518-029] Adapt to new GNATprove_Mode without expansion
As a result of the frontend not doing expansion in GNATprove mode, the
tree has changed, and it is more convenient to perform limited rewritings
in gnat2why instead of the frontend. This is the case here for the rewrite
of calls to renamings or inherited operations in the underlying renamed
or inherited subprogram.

Also, aggregates using box notation are not always expanded anymore into
individual component associations. Mark these as not supported until
further investigation.

* spark/spark_rewrite.ads|s
New units for rewriting of SPARK code before analysis.
* spark/spark_definition.adb
(Mark): Mark partially initialized aggregates not yet supported.
* why/gnat2why-driver.adb
(Rewrite_All_Compilation_Units): New procedure to rewrite all compilation
  units.