[M513-016] flow analysis - stop abuse of tags
The message tags were/are abused to uniquely identify error messages
so that GPS can display the trace files. For flow analysis this was
not sufficient so I abused the system further to embed a variable name
in the tag.

The correct solution will be to use python dictionaries instead of error
messages if gnatprove is run in an "IDE mode".