[M514-037] flow - enable by default
More trouble with printing that particular error: now we check if we
can easily print it, and if not just have a slightly more generic
error (that does not mention the actual parameter). This is OK
because we still point to the correct place.

Also, we ignore "ineffective statement" errors for in parameters as
they are not really very helpful.