[M617-003] flow - signal no-error analysis via file
We now create a file with the same naming convention as for the
dot and pdf files that ends in .flow_ok and contains the string
OK per analysed unit without errors/warnings.

The content of the .flow_ok files is not yet finalized. We might
decide that we need more info to be carried inside.