[K831-027] handle function actuals with subtype of expected type
When a generic has a formal parameter that is a function, it may
be the case that its actual parameter does not have the expected
type but rather some compatible type.
Declare a new why function with the expected signature for the
parameter and use this one in the textual replacement instead.