M821-023 generate a single standalone Why file for each unit
Instead of generating declarations only for the entities in the current
unit, and let Why3 gather the pieces, we generate for each unit a
standalone Why file. Right now this will lead to longer running times of
gnat2why, but removes many constraints on generation of Why.

The notion of Why "file" remains in gnat2why, but it now corresponds to
sections of the single large file.

The Why file is standalone in the sense that it only needs the Why and
GNATprove standard library, but no pregenerated Why files and no files
that come from other Ada units.

spark_definition.adb:
  Add external entities to special list Withed_Entities

backend.adb:
  (Call_Backend) Standard file is now translated always, so no extra
    handling for Standard_Mode

gnat2why-driver.adb:
  (Print_Why_File) print all "files" together, into a single file on
    disk
  (GNAT_To_Why) always translate the standard package first
  (Translate_CUnit) also translate entities from other files
  (Translate_Standard_Package) remove code that was only used in
    "standard_mode"

why-inter.adb:
  (Add_With_Clause, Add_With_For_Entity) never reference an external
    file, always point to modules in the same file, excpept for fixed
    packages such as "int.Int"
  (Dispatch_Entity) no difference between spec and body anymore, so
    regroup lots of cases
  (Init_Why_Files) only create a single file, but stil initialize the
    sections