[M712-013] verification rule that OUT cannot be read
Added this rule and also added that attributes
that depend on the value of an out mode parameter
cannot have attributes that depend on the value
of the object applied to it such as X'Old.
The same rule regarding attributes has been
added to global_items with a mode selector of
Output.