[M830-022] proof - missing theory import for contract cases
To check that at most one of the contract cases is true VCG produces
something like this:

   (if A then (1) else (0)) + (if B then (1) else (0) ... <= 1

The <= 1 obviously uses an infix operator only defined in int.Int.

VCG will cleverly look at entities in the tree and only include
imports that are needed. If you have a contract cases that never uses
integers then this is not included. We now include int.Int in all
precondition VCs.