M821-023 fix order of definitions for entities
It is incorrect to first deal with all with'ed entities, and then with
the entities of the current unit, because there may be dependencies from
the with'ed specs to our spec. This is now fixed.

Also, remove the completions that were specially generated for
expression functions of other units. Not needed anymore.

spark_definition.adb:
   (Append_Entity_To_List, Append_Subprogram_Entities_To_List):
     entities except bodies are appended to spec_entities
   (Mark_Subprogram_Body) expr fun bodies of other units are still of
     interest, because we need to generate axioms for them

gnat2why-driver.adb:
   (Translate_CUnit) No more withed entities, and no more completions
     needed
   (Translate_Entity) Do not generate VCs for subprograms of other units
     (e.g. expression functions in specs)