M821-023 no lemmas in container theories
Change lemmas to axioms in all container theories

When why3 is run on a given file, it tries to prove all lemmas in it. We
do not want to prove the lemmas in the container theory, so we change
them to axioms.

This worked before because lemmas are only proved in the current file,
not in included files, and previously lemmas were only in included
files.