M821-023 rename main why file to "unit.mlw"
previously was called "unit__package.mlw".

This requires some care actually, to avoid clash between the stdlib file
"bool.why" and a file "bool.mlw" that would be produced from a user unit
"bool.adb". Our solution is to *not* add the object directories to the
Why search path. Only a single file is read there at a time, so it's not
needed.

gnat2why-driver.adb
   (Print_Why_File, Touch_Main_File): do not add __package suffix

gnatprove.adb:
    Don't add the object directories to the Why3 search path