[M811-008, M808-025] Control warnings with switch
We now create a default test.gpr files for all tests that do not require
a specific one, and so this patch removes many such files that are not
specific. The default test.gpr sets the -gnatws switch for not generating
frontend warnings. The prove/do_flow script commands are also modified
to set --warnings=on by default, so flow analysis warnings are issued
but not considered as errors (which is the default otherwise), unless
overwritten with the optional arguments passed in test.py