[M905-048] Adapt testsuite to SPARK_Mode
Now, only code that is marked SPARK_Mode => On is analyzed by GNATprove.
For tests without a specific test.gpr, the default test.gpr uses a
local configuration pragma file that sets it.
For tests with their own test.gpr, either this one sets the local
configuration pragma, or (in the majority of cases) SPARK_Mode pragmas
are added in the code as necessary.