[M618-009] Initial support for proof of elaboration code
We now generate VCs for the code of elaboration of a package, and
we attempt to prove that it fulfills the Initial_Condition aspect
if one is specified. This is not complete, as we're not considering
yet the possible effects in local package elaborations.

* gnat2why/spark/spark_util.adb|s
(Get_Package_Body): New function that returns the body of a package.
* gnat2why/why/gnat2why-driver.adb
(Do_Generate_VCs): Generate VCs now also for package that are fully
  in SPARK.
* gnat2why/why/gnat2why-expr.adb
(Transform_Declaration): Get through private base type, and ignore
  private declarations.
* gnat2why/why/gnat2why-nodes.ads
(Subp_Location): Accept packages for parameter.
* gnat2why/why/gnat2why-subprograms.adb|s
(Generate_VCs_For_Package_Elaboration): New procedure to generate
  VCs for a package elaboration.
* gnat2why/why/why-gen-expr.ads
(Cur_Subp_Sloc, Cur_Subp_Name_Label): Update comment to mention that
  these accept also a package.
* common/vc_kinds.ads
(VC_Kind): Add new kind of VC for initial condition of package.