[M314-014] flow - packages
Candidate fix for supporting Initializes => null. I have basically copied
the code from Parse_Dependency_Maps and changed it to search for an
output equal to the one we're looking for.

This is more or less what the code did before, but at least now it works
for the null case and they are very similar.

Ideally this can be re-factored somehow, but the only good idea I had so
far involves writing an iterator for the dependency/initialises aspect
and given how much fun the graph iterators were this is fairly low
priority for now.