M409-009 fix unknown ranges in preconditions
Previously we generated two separate Why functions for each Ada
subprogram: one to prove checks in the precondition and certain
properties of contract cases, and one for all the rest.

The compiler generates at the start of a subprogram body some subtype
declarations, which are sometimes necessary to prove checks in the the
precondition. Such checks were previously unprovable, because we
translated these subtype declarations only in the Why function for the
body.

We now generate a single Why function for all checks, so the above
problem disappears. Still, special care must be taken so that these
compiler-generated declarations are translated before checks for the
precondition are proved.

gnat2why-driver.adb:
  (Generate_VCs) only a single subprogram needs to be called now
gnat2why-expr.adb:
  (Transform_Declarations_From_Source)
  (Transform_Declarations_Not_From_Source)
    Two subprograms to translate only specific parts of the declaration
    list of the subprogram
gnat2why-subprograms.adb:
  (Generate_VCs_For_Subprogram) new function to merge previously
    existing functions Generate_VCs_For_Subprogram_Spec and
    Generate_VCs_For_Subprogram_Body
gnat2why.ads:
  enumerations Generate_VCs_For_Pre and Generate_VCs_For_Post were in
  fact equivalent, merged to Generate_VCs_For_Contract