M926-012 fix record conversion on assignment
We cannot use the expected Why type for generating the discriminant
check, it does not contain all the information. However, the Ada type
usuall contains the correct and useful information. Here we use this to
generate first the correct check, and then to convert to the expected
Why type.

gnat2why-expr.adb
   (Transform_Assignment): go through the Ada type for check, then
     convert to the Why type for typing