[M930-042] Issue different error message for not-yet-supported features
Error message for not-yet-supported features that are in SPARK 2014 is
now different from violation of SPARK rules. It explicitly says "not yet
supported", for example:

   arr.ads:4:09: type "A" of dimension 6 is not yet supported

Pragmas which are not recognized get a warning instead of an error.

* common/spark_violations.adb|s
File removed.
* common/string_utils.ads
Space.
* common/vc_kinds.ads
(SPARK_Suffix): Put string here.
* gnat2why/spark/spark_definition.adb
(Mark): Issue different message for not-yet-supported features.
* gnat2why/why/gnat2why-expr.adb
Add comment.
* gnatprove/spark_report.adb
Adapt to moved SPARK_Suffix.