[M613-039] Accept noop node in actions inserted by the frontend
The frontend may insert actions in expressions, which are a block of
declarations inside the expression. Accept freeze nodes which have no
impact on verification. Deal specially with constants introduced by
the compiler, which may be initialized with an expression that cannot
be translated as a term in an axiom for the constant. Choose not to
generate an axiom for these constants instead. Their value should be
available when translating the corresponding expression as a program.

* gnat2why/spark/spark_definition.adb
(Mark_Actions): Accept freeze-entity nodes as possible actions.
* gnat2why/why/gnat2why-decls.adb
(Translate_Constant_Value): Do not attempt to generate an axiom
  for constants introduced by the compiler.
* gnat2why/why/gnat2why-expr.adb
(Transform_Actions, Transform_Actions_Preparation): Do nothing for
  freeze-entity nodes.