[M926-024] flow - wrong errors on generated effects
I have now enabled yannicks patch and thus enabled flow analysis on the
tests identified in this ticket.

We now also ignore unused *generated* globals which should cut down on
the spam as well (these represent globals that should be proof_in).

Disable flow for the openECTS* tests again as there are new issues
related to handling of main programs.