[M607-025] Further simplify the inclusion between Why theories
Procedure Close_Theory in why-inter.adb is responsible for getting all
necessary includes for the current theory. Until now, this required
2 entities and 3 boolean flags to be passed in parameter to Close_Theory,
which has been simplified to one entity and one "kind" of theory being
closed. This relies on identifying these different kinds of theories we
generate:

  Standalone_Theory      --  standalone definition of symbols
  Definition_Theory      --  definition of symbols
  Axiom_Theory           --  axioms for previously defined symbols
  VC_Generation_Theory   --  generation of VCs

It remains to check whether we can simplify also the handling of theories
for which there is an axiomatization (see Complete_Package_With_External_
Axioms_Translation in gnat2why-decls.adb.)

* gnat2why/why/gnat2why-decls.adb
* gnat2why/why/gnat2why-driver.adb
* gnat2why/why/gnat2why-expr.adb
* gnat2why/why/gnat2why-subprograms.adb
* gnat2why/why/gnat2why-types.adb
Adapt to new signature of Close_Theory.
* gnat2why/why/gnat2why-nodes.adb|s
(Get_Graph_Closure): New version of the function which starts from a set
  of nodes instead of a single node.
* gnat2why/why/why-inter.adb|s
(Close_Theory): Simplify function.