[MA02-028] Always mark specs in SPARK or not
While bodies that are not marked by the user with SPARK_Mode (On) can
be skipped in marking, specs should always be marked as in/out of
SPARK, as this is the basis for deciding whether the entities can
be used from other code.

* gnat2why/spark/spark_definition.adb
(Mark_Package_Declaration, Mark_Subprogram_Declaration): Always mark
the declaration, even if SPARK_Mode is Off.