[M716-042] No assumptions dynamic for constants and types in loops
The treatment of loops in gnat2why unrolls the first part of the
loop before the loop invariant, to create a Why loop around the
loop invariant. This duplicates the code between the start of the
loop and the loop invariant. If this code happens to include
definition of dynamic constants and types (even implicit types
introduced by the compiler), then assumptions added by gnat2why
were inconsistent.

This is not completely sufficient, as the translation of types
is currently based on constants for first/last, that should be made
mutable in Why. To be done in another patch.

* gnat2why/why/gnat2why-expr.adb
(Assume_Of_Scalar_Subtype): Do not introduce an assumption for
  types locally defined in loops.
* gnat2why/why/gnat2why-util.adb|s
(Is_Locally_Defined_In_Loop): New function that returns if node N
  is defined locally in a loop.
(Is_Mutable_In_Why): Return False for constants defined locally
  in loops.