[M318-028] flow - main program
The check for uninitialised variables is now a warning (even if
all these should really be errors). This allows proof of main
programs that use computed globals.

Remove check for ineffective initialisation of globals on the main
program; as initialised globals might be useful in the elaboration
of *other* packages.