[MA15-023] Do not output duplicate messages in IDE mode
When called from within an IDE (with switch --ide-progress-bar),
gnatprove should not issue duplicate messages, which it does currently.
Also, warning messages should start with "warning :". This patch fixes
both issues, but it remains to stop gnatprove in error if an error is
detected (or a warning when treated as errors).

* gnat2why/flow/flow_error_messages.adb
(Print_JSON_Or_Normal_Msg): Issue either a JSON or a regular message.
  Prepend warning string to warning messages.