[MA03-051] Prepare work for removing generated constrained types
Constrained implicit types generated by the frontend may lead to
unsound handling in Why, so they should not be generated in
GNATprove_Mode. This is a first step for achieving this goal.
The next step is to disable Expand_Subtype_From_Expr in exp_util.adb
in GNATprove_Mode, and have a consistent handling of unconstrained
types in gnat2why.

* gnat2why/spark/spark_util.adb|s
(MUT_Kind): New function to get kind of most-underlying-type.
(Has_.._Type): Use MUT_Kind instead of MUT, so that it works also for
  types in units with external axiomatization.
* gnat2why/why/gnat2why-decls.adb
(Translate_Variable): Only parameters should be translated as split.
* gnat2why/why/why-gen-expr.adb
(Insert_Record_Conversion): Do not use root type for types in units
  with external axiomatization.
* gnat2why/why/why-inter.adb
(Base_Why_Type): Do not use root type for types in units
  with external axiomatization.