[MA14-023] Improve example of Loop_Entry in 3.3.2
The Move example had flow errors which were unacceptable
and it was a little non-intuitive, since the same thing could be
achieved with a simple aggregate assignment.

This new example, from TJJ, illustrates the use of 'Loop_Entry
in a procedure that reverses the contents of an array in-place.
Flow analysis and Proof both 100% OK now.