[MA03-034] Add minimal support for attribute Valid
Attribute Valid is supported a minima, by assuming it always evaluates to
True. Its prefix is analyzed for run-time errors. A warning is issued to
notify users that Valid is analyzed this way.

* gnat2why/spark/spark_definition.adb
(Mark_Attribute_Reference): Issue warning on Valid instead of error.
* gnat2why/why/gnat2why-expr.adb
(Transform_Attr): Add minimal support for Valid.