[MB01-007] Overflow mode should only impact integer operations
Pragma Overflow_Mode (or the equivalent switch -gnato??) should only
impact the generation of code and the verification of expressions of
signed integer type. Other expressions (of enumerated type, or real type)
which carry overflow checks should not be impacted. This was not the
case, and it is fixed by this patch.

* gnat2why/why/gnat2why-expr.adb
(Transform_Expr): Use overflow mode only on expressions of signed integer
  type.