[MB05-010] Use ordered sets of nodes in gnat2why
Set of nodes used in gnat2why may impact the order in which entities
are entities are declared in the generated VCs, which may in turn
impact the time to prove a VC, due to heuristics in the prover.
Use ordered sets to avoid variability between machines with hashed
sets.

* gnat2why/why/gnat2why-decls.adb
(Translate_Variable): Theory for a variable may need to include
  modules that define types. Make it a definition theory.
* gnat2why/why/gnat2why-nodes.ads
(Node_Sets): Change from hashed to ordered sets.
* gnat2why/why/why-inter.adb
(Add_Use_For_Entity): Correct typo that caused all theories to be
  imported.