[MB14-016] Include package without body in report file
A package without body (either because no body is needed, or because it
is not available) is analyzed by GNATprove for elaboration, hence it
should have an entry in the report file. This corrects in particular
the case where such packages are the only units analyzed, which caused
previously GNATprove to issue the message that no code was analyzed.

* gnat2why/spark/spark_definition.adb
(Generate_Output_In_Out_SPARK): Generate value "all" for package in SPARK
  without a body. This value replaces the previous value "body".
(Mark_Package_Declaration): Generate entry in log file for packages from
  the current unit for which no body is available.
* gnatprove/spark_report.adb
(Handle_SPARK_File): Use new value "all" instead of "body".