[MA29-016] flow - refinement
Make use of the new refinement facilities - we should now always use the
appropriate (refined_)global and (refined_)depends, depending on the
current scope.

In addition, if a (refined_)global or (refined_)depends is not refined
enough, we magically expand it. This can happen for nested packages or
for expression functions implemented in the spec or private part.

The Get_Global infrastructure has changed, we now have smaller wrapper
functions for proof which return the most refined view possible.

Smaller things fixed:
   - For abstract state entities, we now print Package.State instead of
     just State (which should make errors involving abstract state easier
     to read)
   - Fix a bug in Get_Globals where Proof_Ins did not consider
     Globals_For_Proof
   - Quote the correct contract when complaining about missing variables
     (used to be always "Global", it is now "Refined_Global" if it exists)
   - More documentation in Flow_Refinement

We still need to fix analysis of pre- and postconditions and contract
cases. They may sometimes be analysed in the incorrect scope and the
resulting error messages will make no sense.